requirements/base.txt
Django==1.11.6
django-filter==2.2.0
requests==2.21.0
djangorestframework==3.10.3
pandas==0.25.3
